books-service/requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
drf-yasg==1.21.8
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
gunicorn==23.0.0
requests==2.32.3
